Tuesday Night Update
I've been struggling to implement the turning system I need. Specific problems are outlined in the code. I've roadmapped most of the other functions I need, but without this working, it's difficult to implement them. I don't anticipate any other aspect of this project being as complicated as this has been, so once I get past this roadblock, everything else should be fine. If you're in Santa Cruz this Friday I'm definitely coming to office hours.